// src/app/shared/model/token-transfer.model.ts
export class TokenTransferModel {
    balance: number = 0;
    amount: number = 88888888888888;
    transaction_fee: number = 0;
    wallet_address: string = '';
    symbol: string = '';
    logo: string = '';
}
export class TokenModel{
    logo: string = '';
    token_name: string = '';
    symbol: string = '';
    balance: number = 0;
}